Update courseraCheck.R
Is it possible to change the email stored in an instance of swirl if it was incorrectly entered? It just so happens that at the moment that question appeared on my screen I was distracted. How can I change it? 

Maybe I can add a confirm pass word script for you all to confirm. That's all I am trying to do.